requirements.txt
accelerate~=0.27.2
CairoSVG~=2.7.1
datasets~=2.17.0
diffusers~=0.26.3
huggingface_hub~=0.20.3
numpy~=1.26.4
pandas~=2.2.1
peft~=0.8.2
Pillow~=10.2.0
pyarrow~=15.0.0
torch~=2.2.0
torchvision~=0.17.0
tqdm~=4.66.2
transformers~=4.38.1
xformers~=0.0.24
xmltodict~=0.13.0
